decouple node help between sidebar and right panel
  - Extract useNodeHelpContent composable so NodeHelpContent fetches its own
    content, allowing multiple panels to show help independently
  - Remove sync behavior from NodeHelpPage that caused left sidebar to change
    when selecting different graph nodes
  - Add telemetry tracking for node library help button
